Naive paramgen: Ensure all IDs are multiplied by int64 values (and not int32s which may cause overflow

+++ naive-paramgen/paramgen-queries/pg-20.sql
@@ -4,5 +4,5 @@ SELECT
 FROM
     companyNumEmployees,
     (SELECT personId FROM personNumFriends ORDER BY personId LIMIT 100)
-ORDER BY md5(3532569367*companyId + 211*personId)
+ORDER BY md5(3532569367::bigint*companyId + 211::bigint*personId)
 LIMIT 400
