+++ naive-paramgen/paramgen-queries/pg-20b.sql
@@ -4,5 +4,5 @@ SELECT
 FROM
     companyNumEmployees,
     (SELECT id FROM personNumFriends ORDER BY md5(id) LIMIT 100) personNumFriends_sample
-ORDER BY md5(3532569367*companyNumEmployees.id + 211*personNumFriends_sample.id)
+ORDER BY md5(3532569367::bigint*companyNumEmployees.id + 211::bigint*personNumFriends_sample.id)
 LIMIT 400
